moved app to components/pages
